+++ src/game/tests.rs
@@ -542,6 +542,7 @@ mod vintage_cube_talisman_of_progress;
 mod vintage_cube_talon_gates_of_madara;
 mod vintage_cube_tamiyo;
 mod vintage_cube_tamiyo_inquisitive_student;
+mod vintage_cube_tarfire;
 mod vintage_cube_tear_asunder;
 mod vintage_cube_teferi_time_raveler;
 mod vintage_cube_tendrils;